feat: add tools to read NBF files
This commit primarily adds the ability to read NBF files to
PegasusTools. This is done through the public `load_file` function
which determines the file type then calls the appropriate loading
function. Currently that is only the function for loading NBF files
but more file types will be supported soon. The new functions are
15-22x faster to load NBF files than the original ones used by the
group.

In addition there are a handful of small, semi-related, changes:

- Add documentation for the new features
- Add tests for all new functions
- Remove files used to test the packaging process
- Add numpy and matplotlib dependenies
- Modify pytest options to always show coverage
- Minor tweaks to Ruff linter settings
- Add some repo specific files to the gitignore