refactor!: system for loading NBF files
This change primarily removes `load_file` and integrates `_load_nbf` and
the functions it calls into the constructor for `PegasusNBFData`. This
was done because implementing `load_file` as a wrapper over all
different loading functions was becoming too complex and was generally
unneccarsy and confusing with minimal upside. Integrating `_load_nbf`
into the constructor of `PegasusNBFData` was done because its purpose
was to construct a `PegasusNBFData` object so it should be the
constructor. Now NBF files can be read by just instantiating a
`PegasusNBFData` object.

Additionally:

- integrate `nbf_testing_utilities.py` into `test_loading_nbf.py`
- Update NBF tests for new version of `PegasusNBFData`
- Update documentation for loading NBF files to remove `load_file`
  and replace it with `PegasusNBFData`

BREAKING CHANGE: Remove `load_file`. Reading NBF files should be done
with the new version of `PegasusNBFData`.